# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoraPay API",
    "version": "0.1.0",
    "description": "Cross-border remittance platform. Russia/CIS → Nigeria & Ghana.\n\n**No live funds.** `LIVE_FUNDS_ENABLED` defaults to false and every provider\nruns against a simulator until partner agreements and written legal opinions\nare in place.\n\nTwo conventions that differ from most payment APIs, and both are deliberate:\n\n1. **Money is never a JSON number.** Amounts carry `minorUnits` as a string,\n   which is the authoritative field. A JSON number is a double, and a double\n   is not a currency.\n2. **An acknowledgement is not a settlement.** A provider that accepts a\n   submission has moved no money. Only a status poll or the T+1 statement\n   produces an outcome, and only an outcome posts to the ledger.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.staging.morapay.example",
      "description": "Staging"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Back-office tokens are signed with a different key from customer tokens. A customer token fails signature verification here, not merely authorisation."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/admin/auth/login": {
      "post": {
        "tags": [
          "Back office"
        ],
        "summary": "Staff sign in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresInSeconds": {
                      "type": "integer"
                    },
                    "staff": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "SUPPORT",
                              "COMPLIANCE_OFFICER",
                              "TREASURY_OPERATOR",
                              "ADMIN"
                            ]
                          }
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "roles"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresInSeconds",
                    "staff"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/compliance/cases": {
      "get": {
        "tags": [
          "Back office"
        ],
        "summary": "Compliance review queue",
        "description": "Requires one of: COMPLIANCE_OFFICER, SUPPORT, ADMIN.",
        "security": [
          {
            "adminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/transfers": {
      "get": {
        "tags": [
          "Back office"
        ],
        "summary": "Operational transfer search",
        "description": "Requires one of: SUPPORT, COMPLIANCE_OFFICER, ADMIN, TREASURY_OPERATOR.",
        "security": [
          {
            "adminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/treasury/positions": {
      "get": {
        "tags": [
          "Back office"
        ],
        "summary": "Float positions and open FX exposure",
        "description": "Requires one of: TREASURY_OPERATOR, ADMIN.",
        "security": [
          {
            "adminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/treasury/prefunding": {
      "get": {
        "tags": [
          "Back office"
        ],
        "summary": "Prefunding requests and approvals",
        "description": "Requires one of: TREASURY_OPERATOR, ADMIN.",
        "security": [
          {
            "adminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}